requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
sqlparse==0.4.4
typing_extensions==4.8.0